feat: 게시글 좋아요 수정
 좋아요 누르거나 삭제시, 좋아요 수 가져옴

Resolves:
Related to: S11P12D208-277

+++ backend/src/main/java/com/campforest/backend/board/controller/BoardController.java
@@ -153,7 +153,7 @@ public class BoardController {
 				return ApiResponse.createError(ErrorCode.LIKE_ALREADY_EXISTS);
 			}
 			boardService.likeBoard(boardId, userId);
-			return ApiResponse.createSuccessWithNoContent("게시글 좋아요 성공하였습니다");
+			return ApiResponse.createSuccess(boardService.getBoard(boardId).getLikeCount(),"게시글 좋아요 성공하였습니다");
 		} catch (Exception e) {
 			log.error("Like operation failed", e);
 			return ApiResponse.createError(ErrorCode.INTERNAL_SERVER_ERROR);
@@ -163,13 +163,14 @@ public class BoardController {
 	//게시글 좋아요 삭제
 	@DeleteMapping("like")
 	public ApiResponse<?> deleteLike(@RequestParam Long boardId, @RequestParam Long userId) {
-		try {
+
+			if (!boardService.checkLike(boardId, userId)) {
+				return ApiResponse.createError(ErrorCode.LIKE_NOT_FOUND);
+			}
 			boardService.deleteLike(boardId, userId);
-			return ApiResponse.createSuccessWithNoContent("게시글 좋아요 삭제 성공하였습니다");
-		} catch (Exception e) {
-			log.error("Failed to delete like", e);
-			return ApiResponse.createError(ErrorCode.LIKE_NOT_FOUND);
-		}
+
+			return ApiResponse.createSuccess(boardService.getBoard(boardId).getLikeCount(),"게시글 좋아요 삭제 성공하였습니다");
+
 	}
 
 	//게시글별 좋아요 갯수 조회
